# feegow-challenge-covid-backend/storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "API de Vacinação de Funcionários",
        "description": "Uma API para registro de vacinação de funcionários contra a COVID-19.",
        "contact": {
            "name": "Suporte Técnico",
            "email": "suporte@empresa.com"
        },
        "version": "1.0.0"
    },
    "paths": {
        "/api/funcionarios": {
            "get": {
                "tags": [
                    "Funcionarios"
                ],
                "summary": "Lista todos os funcionários",
                "operationId": "ea699d46fc67c0fbfd4a5b47e692c7bd",
                "responses": {
                    "200": {
                        "description": "Operação bem-sucedida"
                    },
                    "404": {
                        "description": "Nenhum Funcionário encontrado."
                    }
                }
            },
            "post": {
                "tags": [
                    "Funcionarios"
                ],
                "summary": "Cria um novo funcionário",
                "operationId": "1243a4b3afb8f96b38e6d64d1830072e",
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "required": [
                                    "cpf",
                                    "nome_completo",
                                    "data_nascimento",
                                    "portador_comorbidade"
                                ],
                                "properties": {
                                    "cpf": {
                                        "type": "string",
                                        "format": "regex:/^\\d{11}$/",
                                        "example": "12345678900"
                                    },
                                    "nome_completo": {
                                        "type": "string",
                                        "example": "João Silva"
                                    },
                                    "data_nascimento": {
                                        "type": "string",
                                        "format": "date",
                                        "example": "1990-01-01"
                                    },
                                    "portador_comorbidade": {
                                        "type": "boolean",
                                        "example": true
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "201": {
                        "description": "Funcionário criado com sucesso"
                    },
                    "404": {
                        "description": "Funcionário não encontrado."
                    }
                }
            }
        },
        "/api/funcionarios/{id}": {
            "get": {
                "tags": [
                    "Funcionarios"
                ],
                "summary": "Exibe um funcionário específico",
                "operationId": "37d4fbf9d87bd6edf167e2014fbdcccc",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Operação bem-sucedida"
                    },
                    "404": {
                        "description": "Funcionário não encontrado."
                    }
                }
            },
            "delete": {
                "tags": [
                    "Funcionarios"
                ],
                "summary": "Deleta um funcionário específico",
                "operationId": "68308abb641f5b9baadbbca03fbb39d9",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "204": {
                        "description": "Funcionário deletado com sucesso"
                    },
                    "404": {
                        "description": "Funcionário não encontrado."
                    }
                }
            }
        },
        "/api/funcionarios/{funcionario}": {
            "put": {
                "tags": [
                    "Funcionarios"
                ],
                "summary": "Atualiza um funcionário específico",
                "operationId": "8539ecc3d84efbd05807d375bb561207",
                "parameters": [
                    {
                        "name": "funcionario",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "required": [
                                    "cpf",
                                    "nome_completo",
                                    "data_nascimento",
                                    "portador_comorbidade"
                                ],
                                "properties": {
                                    "cpf": {
                                        "type": "string",
                                        "format": "regex:/^\\d{11}$/",
                                        "example": "12345678900"
                                    },
                                    "nome_completo": {
                                        "type": "string",
                                        "example": "João Silva"
                                    },
                                    "data_nascimento": {
                                        "type": "string",
                                        "format": "date",
                                        "example": "1990-01-01"
                                    },
                                    "portador_comorbidade": {
                                        "type": "boolean",
                                        "example": true
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Funcionário atualizado com sucesso"
                    },
                    "404": {
                        "description": "Funcionário não encontrado."
                    }
                }
            }
        },
...
        "/funcionario-vacina/{funcionario_cpf}/{vacina_id}/{dose}": {
            "put": {
                "tags": [
                    "FuncionarioVacina"
                ],
                "summary": "Atualiza os detalhes de uma vacinação específica de um funcionário",
                "operationId": "update",
                "parameters": [
                    {
                        "name": "funcionario_cpf",
                        "in": "path",
                        "description": "CPF do funcionário",
                        "required": true,
                        "schema": {
                            "type": "string"
                        }
                    },
                    {
                        "name": "vacina_id",
                        "in": "path",
                        "description": "ID da vacina",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    },
                    {
                        "name": "dose",
                        "in": "path",
                        "description": "Dose da vacina",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "required": [
                                    "data_dose"
                                ],
                                "properties": {
                                    "data_dose": {
                                        "type": "string",
                                        "format": "date",
                                        "example": "2024-03-14"
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Registro de vacinação atualizado com sucesso!"
                    }
                }
            },
            "delete": {
                "tags": [
                    "FuncionarioVacina"
                ],
                "summary": "Remove uma vacinação específica de um funcionário",
                "operationId": "destroy",
                "parameters": [
                    {
                        "name": "funcionario_cpf",
                        "in": "path",
                        "description": "CPF do funcionário",
                        "required": true,
                        "schema": {
                            "type": "string"
                        }
                    },
                    {
                        "name": "vacina_id",
                        "in": "path",
                        "description": "ID da vacina",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    },
                    {
                        "name": "dose",
                        "in": "path",
                        "description": "Dose da vacina",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "204": {
                        "description": "Registro de vacinação removido com sucesso!"
                    }
                }
            }
        },
...
    }
}